incompleto. aggiunte classi, modificate interfacce
