--- src/VoiceStyle.java
@@ -1,26 +1,0 @@
-public class VoiceStyle {
-
-    private String voiceType;
-    private String voicePitch;
-
-    public VoiceStyle(String voiceType, String voicePitch) {
-        this.voiceType = voiceType;
-        this.voicePitch = voicePitch;
-    }
-
-    public String getVoiceType() {
-        return voiceType;
-    }
-
-    public void setVoiceType(String voiceType) {
-        this.voiceType = voiceType;
-    }
-
-    public String getVoicePitch() {
-        return voicePitch;
-    }
-
-    public void setVoicePitch(String voicePitch) {
-        this.voicePitch = voicePitch;
-    }
-}
