+++ src/it/project1/monster/Ability.java
@@ -1,4 +1,4 @@
-package monster;
+package it.project1.monster;
 
 public interface Ability {
     String getName();
