+++ src/it/project1/monster/Condor.java
@@ -1,7 +1,7 @@
-package monster;
+package it.project1.monster;
 
-public class Eagle extends Monster implements FlyingMonster {
-    public Eagle(String name, int level, int monsterHealtPoints, int monsterAbilityPoints, int monsterAttack, int monsterDefense) {
+public class Condor extends Monster implements FlyingMonster {
+    public Condor(String name, int level, int monsterHealtPoints, int monsterAbilityPoints, int monsterAttack, int monsterDefense) {
         super(name, level, monsterHealtPoints, monsterAbilityPoints, monsterAttack, monsterDefense);
     }
