+++ src/it/project1/monster/Eagle.java
@@ -1,9 +1,10 @@
-package monster;
+package it.project1.monster;
 
-public class Crow extends Monster implements FlyingMonster {
-    public Crow(String name, int level, int monsterHealtPoints, int monsterAbilityPoints, int monsterAttack, int monsterDefense) {
+public class Eagle extends Monster implements FlyingMonster {
+    public Eagle(String name, int level, int monsterHealtPoints, int monsterAbilityPoints, int monsterAttack, int monsterDefense) {
         super(name, level, monsterHealtPoints, monsterAbilityPoints, monsterAttack, monsterDefense);
     }
+
     @Override
     public String getType() {
         return "Flying";
