+++ src/it/project1/monster/Element.java
@@ -1,7 +1,7 @@
 /**
  * @author Figuccia
  */
-package monster;
+package it.project1.monster;
 
 public abstract class Element {
     private final String name;
