+++ src/it/project1/monster/FireElement.java
@@ -1,4 +1,4 @@
-package monster;
+package it.project1.monster;
 
 public class FireElement extends Element {
     public FireElement() {
