+++ src/it/project1/monster/MonsterAbilities.java
@@ -1,7 +1,7 @@
 /**
  * @author Figuccia
  */
-package monster;
+package it.project1.monster;
 
 import java.util.List;
