+++ src/it/project1/monster/MonsterClass.java
@@ -1,4 +1,4 @@
-package monster;
+package it.project1.monster;
 
 public class MonsterClass {
 }
