+++ src/it/project1/monster/NormalMonster.java
@@ -1,4 +1,4 @@
-package monster;
+package it.project1.monster;
 
 public interface NormalMonster {
 }
