+++ src/it/project1/monster/Rideable.java
@@ -1,7 +1,7 @@
 /**
  * @author Figuccia
  */
-package monster;
+package it.project1.monster;
 
 public interface Rideable {
     void ride();
